(#14) develop

+++ app/src/main/java/com/example/ignis/main/SearchResponse.kt
@@ -1,16 +1,16 @@
 package com.example.ignis.main
 data class SearchResponse(
-    val count: Int,
+    val count: Long,
     val feeds: List<Feed>
 )
 
 data class Feed(
-    val id: Int,
+    val id: Long,
     val title: String,
     val imageUrl: String,
     val user: String,
     val createAt: String,
-    val count: Int,
-    val x: Int,
-    val y: Int
+    val count: Long,
+    val x: Double,
+    val y: Double
 )
